refactor: split non-ts artifact templates
