+++ .sampo/changesets/1006-split-non-ts-artifact-templates.md
@@ -1,0 +1,6 @@
+---
+npm/@wp-typia/project-tools: patch
+npm/wp-typia: patch
+---
+
+Split compound and persistence non-TypeScript scaffold templates into focused runtime modules without changing generated output.
